courseinfo step9 + seperate module
