+++ part2/courseinfo/src/App.jsx
@@ -1,57 +1,59 @@
-const Course = ({ course }) => {
-  return (
-    <div>
-      <Header courseName={course.name} />
-      <Content courseParts={course.parts} />
-      <Total sum={course.parts.reduce((sum, part) => sum + part.exercises, 0)} />
-    </div>
-  )
-}
-
-const Header = ({ courseName }) => <h1>{courseName}</h1>
-
-const Content = ({ courseParts }) =>
-  <div>
-    {courseParts.map(part =>
-      <Part key={part.id} part={part} />
-    )}
-  </div>
-  
-const Part = ({ part }) =>
-  <p>
-    {part.name} {part.exercises}
-  </p>
-
-const Total = ({ sum }) => <b>total of {sum} exercises</b>
+import Course from './components/Course'
 
 const App = () => {
-  const course = {
-    id: 1,
-    name: 'Half Stack application development',
-    parts: [
-      {
-        name: 'Fundamentals of React',
-        exercises: 10,
-        id: 1
-      },
-      {
-        name: 'Using props to pass data',
-        exercises: 7,
-        id: 2
-      },
-      {
-        name: 'State of a component',
-        exercises: 14,
-        id: 3
-      },
-      {
-        name: 'Redux',
-        exercises: 11,
-        id: 4
-      }
-    ]
-  }
-  return <Course course={course} />
+  const courses = [
+    {
+      name: 'Half Stack application development',
+      id: 1,
+      parts: [
+        {
+          name: 'Fundamentals of React',
+          exercises: 10,
+          id: 1
+        },
+        {
+          name: 'Using props to pass data',
+          exercises: 7,
+          id: 2
+        },
+        {
+          name: 'State of a component',
+          exercises: 14,
+          id: 3
+        },
+        {
+          name: 'Redux',
+          exercises: 11,
+          id: 4
+        }
+      ]
+    }, 
+    {
+      name: 'Node.js',
+      id: 2,
+      parts: [
+        {
+          name: 'Routing',
+          exercises: 3,
+          id: 1
+        },
+        {
+          name: 'Middlewares',
+          exercises: 7,
+          id: 2
+        }
+      ]
+    }
+  ]
+
+  return (
+    <div>
+      <h1>Web development curriculum</h1>
+      {courses.map(course =>
+        <Course key={course.id} course={course} />
+      )}
+    </div>
+  )
 }
 
 export default App
